test(xdg): reuse the Unix test module
